% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pluripotent-package.R
\docType{package}
\name{pluripotent-package}
\alias{pluripotent}
\alias{pluripotent-package}
\title{pluripotent: Generate an embryonic repository}
\description{
Repeatedly generate an embryonic repository with one command.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OuhscBbmc/pluripotent}
  \item \url{http://ouhsc.edu/bbmc/}
  \item Report bugs at \url{https://github.com/OuhscBbmc/pluripotent/issues}
}

}
\author{
\strong{Maintainer}: Will Beasley \email{wibeasley@hotmail.com} (\href{https://orcid.org/0000-0002-5613-5006}{ORCID})

Other contributors:
\itemize{
  \item Andriy Koval [contributor]
}

}
\keyword{internal}
