% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populate-config.R
\name{populate_config}
\alias{populate_config}
\title{Populated the configuration value with project-specific values.}
\usage{
populate_config(path_in, project_name, path_out = path_in)
}
\arguments{
\item{path_in}{\link{character} of the config file to populate.  Required.}

\item{project_name}{\link{character} of the new project.  Required.}

\item{path_out}{\link{character} of the config file to populate.  Required.}
}
\description{
\code{populate_config()} starts with config file and replaces the
template values with project-specific values.
}
\examples{
library(pluripotent)

\dontrun{
url_config <- "https://github.com/OuhscBbmc/cdw-skeleton-1/blob/main/config.yml?raw=true"
path_in    <- "~/config.yml"
path_out   <- "~/config-out.yml"
utils::download.file(url=url_config, destfile=path_in)
populate_config(
  path_in        = path_in,
  project_name   = "thumann-awesomeness-1",
  path_out       = path_out
)
}
}
\author{
Will Beasley
}
