% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-file-list.R
\name{retrieve_file_list}
\alias{retrieve_file_list}
\alias{download_file_list}
\title{Retrieve and download files for a project}
\usage{
retrieve_file_list(
  offspring,
  project_name = "new-project",
  destination_directory = file.path("~", project_name)
)
download_file_list(
  offspring,
  project_name  = "new-project",
  destination_directory = file.path("~", project_name)
)
}
\arguments{
\item{offspring}{\link{character} of observed values in the investigation. Required.}

\item{project_name}{\link{character} of the new project.  Defaults to \code{new-project}.}

\item{destination_directory}{\link{character} of the files.  Defaults to \code{~} (\emph{i.e}, the home directory).}
}
\value{
Table of files to copy
}
\description{
\code{retrieve_file_list()} retrieve the complete list
of files to be copied into a new project, while
\code{download_file_list()} copies the files to the \code{destination_directory}
on your local maachine.
}
\details{
Currently, two types of project is supported:
\enumerate{
\item r-analysis-skeleton
\item cdw-sekeleton-1
}
}
\note{
To view the files involved in each project type,
refer to https://github.com/OuhscBbmc/pluripotent/inst/data/file-to-copy.csv.
Each line represents one file that will be placed in the new project.  The columns are
\enumerate{
\item \code{offspring}: the project type to be created.
\item \code{destination}: the file's location in the new project.
\item \code{source}: the location where the file is copied from.
}
}
\examples{
library(pluripotent)
d1 <- retrieve_file_list("r-analysis-skeleton")

\dontrun{
download_file_list("r-analysis-skeleton", "new-project", "./data-public/testing")
}
}
\author{
Will Beasley
}
