% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start-project.R
\name{start_project}
\alias{start_project}
\alias{start_r_analysis_skeleton}
\alias{start_cdw_skeleton_1}
\title{Start specific projects}
\usage{
start_r_analysis_skeleton(
  project_name  = "new-project-analysis",
  destination_directory = "~/analysis"
)
start_cdw_skeleton_1(
  project_name  = "new-project-cdw",
  destination_directory = "~/cdw"
)
}
\arguments{
\item{project_name}{\link{character} of the new project.  Defaults to \code{new-project}.}

\item{destination_directory}{\link{character} of the files.  Defaults to \code{~} (\emph{i.e}, the home directory).}
}
\description{
\code{start_r_analysis_skeleton()} copies the files
to the \code{destination_directory} on your local maachine.
}
\details{
Currently, two types of project is supported:
\enumerate{
\item r-analysis-skeleton
\item cdw-sekeleton-1
}
}
\note{
To view the files involved in each project type,
refer to https://github.com/OuhscBbmc/pluripotent/inst/data/file-to-copy.csv.
Each line represents one file that will be placed in the new project.  The columns are
\enumerate{
\item \code{offspring}: the project type to be created.
\item \code{destination}: the file's location in the new project.
\item \code{source}: the location where the file is copied from.
}
}
\examples{
library(pluripotent)

\dontrun{
start_r_analysis_skeleton("new-project", "./data-public/testing")
start_cdw_skeleton_1()
start_cdw_skeleton_1("thumann-awesomeness-4")
}
}
\author{
Will Beasley
}
