% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve-file-list.R
\name{retrieve_file_list}
\alias{retrieve_file_list}
\title{Retrieve files for a project}
\usage{
retrieve_file_list(offspring, project_name = "new-project")
}
\arguments{
\item{offspring}{\link{character} of observed values in the investigation. Required.}

\item{project_name}{\link{character} of the new project.  Defaults to \code{new-project}.}
}
\value{
Table of files to copy
}
\description{
Retrieve the complete list of files to be copied into a new project.
}
\details{
Currently, only one type of project is supported:
\enumerate{
\item r-analysis-skeleton
}
}
\note{
To view the files involved in each project type,
refer to https://github.com/OuhscBbmc/pluripotent/inst/data/file-to-copy.csv.
Each line represents one file that will be placed in the new project.  The columns are
\enumerate{
\item \code{offspring}: the project type to be created.
\item \code{destination}: the file's location in the new project.
\item \code{source}: the location where the file is copied from.
}
}
\examples{
library(magrittr)
d1 <- retrieve_file_list("r-analysis-skeleton")
d2 <- retrieve_file_list("r-analysis-skeleton")

# download.file(d1$source, d1$destination)
}
\author{
Will Beasley
}
